+++ src/ObjCylinder.cpp
@@ -39,11 +39,11 @@ Cylinder::Cylinder(ifstream &f)
 
 			if(lineContents.size() < 3) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num2 = (double)atof(lineContents.front().c_str());
+			num2 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num3 = (double)atof(lineContents.front().c_str());
+			num3 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			if(word == "origin") this->origin = Point(num1,num2,num3);
@@ -60,7 +60,7 @@ Cylinder::Cylinder(ifstream &f)
 
 			if(lineContents.size() < 1) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			if(word == "radius")			radius = num1;
