+++ src/ObjPlane.cpp
@@ -30,11 +30,11 @@ Plane::Plane(ifstream &f) {
 
 			if(lineContents.size() < 3) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num2 = (double)atof(lineContents.front().c_str());
+			num2 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num3 = (double)atof(lineContents.front().c_str());
+			num3 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			if(word == "origin")		this->origin = Point(num1,num2,num3);
@@ -54,9 +54,9 @@ Plane::Plane(ifstream &f) {
 
 			if(lineContents.size() < 2) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num2 = (double)atof(lineContents.front().c_str());
+			num2 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			this->width = num1;
@@ -69,7 +69,7 @@ Plane::Plane(ifstream &f) {
 
 			if(lineContents.size() < 1) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			if(word == "reflect")			this->material.reflection = num1;
