+++ src/ObjSphere.cpp
@@ -29,11 +29,11 @@ Sphere::Sphere(ifstream &f)
 
 			if(lineContents.size() < 3) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num2 = (double)atof(lineContents.front().c_str());
+			num2 = atof(lineContents.front().c_str());
 			lineContents.pop();
-			num3 = (double)atof(lineContents.front().c_str());
+			num3 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			if(word == "origin")		this->origin = Point(num1,num2,num3);
@@ -46,7 +46,7 @@ Sphere::Sphere(ifstream &f)
 
 			if(lineContents.size() < 1) break;
 
-			num1 = (double)atof(lineContents.front().c_str());
+			num1 = atof(lineContents.front().c_str());
 			lineContents.pop();
 
 			if(word == "radius")			this->radius = num1;
